구조 변경
